# docs/swagger.yaml
basePath: /api
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.Task:
    properties:
      description:
        type: string
      id:
        type: integer
      is_done:
        type: boolean
      title:
        type: string
    required:
    - title
    type: object
  models.TaskUpdateInput:
    properties:
      description:
        type: string
      is_done:
        type: boolean
      title:
        type: string
    type: object
  models.TodoList:
    properties:
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  models.TodoListUpdateInput:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  models.User:
    properties:
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
info:
  contact: {}
  description: Same learn Go project
  title: Todo API
  version: 0.0.1
paths:
  /api/lists:
    get:
      consumes:
      - application/json
      description: '"Get all todo lists"'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.TodoList'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get todo lists
      tags:
      - Lists
    post:
      consumes:
      - application/json
      description: '"Create todo list"'
      parameters:
      - description: todo list data
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/models.TodoList'
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            type: int
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create list
      tags:
      - Lists
    put:
      consumes:
      - application/json
      description: '"Update list"'
      parameters:
      - description: list id
        in: path
        name: id
        required: true
        type: integer
      - description: list update data
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/models.TodoListUpdateInput'
      produces:
      - application/json
      responses:
        "201":
          description: status
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Update list
      tags:
      - Lists
  /api/lists/{id}:
    delete:
      consumes:
      - application/json
      description: '"Delete list"'
      parameters:
      - description: list id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: status
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Delete list
      tags:
      - Lists
    get:
      consumes:
      - application/json
      description: '"Get list by id"'
      parameters:
      - description: list id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TodoList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get list
      tags:
      - Lists
  /api/lists/{id}/tasks:
    get:
      consumes:
      - application/json
      description: '"Get all tasks from todo list"'
      parameters:
      - description: list id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: '"Create task"'
      parameters:
      - description: list id
        in: path
        name: id
        required: true
        type: integer
      - description: task data
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            type: int
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create task
      tags:
      - Tasks
  /api/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: '"Delete task"'
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: status
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Delete task
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: '"Get task by id"'
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: '"Update task"'
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: integer
      - description: task update data
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/models.TaskUpdateInput'
      produces:
      - application/json
      responses:
        "201":
          description: status
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Update task
      tags:
      - Tasks
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login for get jwt token
      parameters:
      - description: credential
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: sign-in
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: register and create new user
      parameters:
      - description: User info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: int
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: sign-up
      tags:
      - Auth
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"